move extra functions to aux3d
